workshop4/news_agent_service/requirements.txt
google-adk
litellm
requests
python-dotenv
uvicorn
fastapi
nltk